requirements.txt
# Smart Contract Security Auditor — Dependencies
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
numpy>=1.24.0
requests>=2.31.0
gradio>=4.0.0
